requirements.txt
fastapi
uvicorn[standard]
aiosqlite
jinja2
python-dotenv
httpx
